# docs/openapi.yaml
openapi: 3.0.3
info:
  title: True Crime Club API
  description: |
    Especificação OpenAPI para a API do True Crime Club. Esta API conecta todas as partes dinâmicas da aplicação (login, loja, checkout, área do cliente) e servirá tanto para o mock interno quanto para o backend externo definitivo.
    
    Convenções gerais:
    * Valores monetários são expressos em centavos (ex: `14990` equivale a `R$ 149,90`).
    * Datas seguem o formato ISO 8601 (`YYYY-MM-DDTHH:mm:ss.sssZ`).
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Servidor Local (Mockado)
  - url: https://api.truecrimeclub.com.br/v1
    description: Servidor de Produção (Futuro)
...
paths:
...
  /auth/login:
    post:
      tags:
        - Autenticação
      summary: Realizar login do cliente
      description: Autentica o cliente e emite um cookie de sessão ou JWT.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - senha
              properties:
                email:
                  type: string
                  format: email
                  example: mariana.silva@email.com
                senha:
                  type: string
                  format: password
                  example: segredoforte
      responses:
        '200':
          description: Autenticação realizada com sucesso.
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: jwt_exemplo
                  cliente:
                    $ref: '#/components/schemas/Cliente'
        '401':
          description: E-mail ou senha inválidos.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
...
  /customer/subscription:
    get:
      tags:
        - Cliente & Painel
      summary: Obter a assinatura ativa do cliente
      responses:
        '200':
          description: Dados da assinatura.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assinatura'
        '404':
          description: Nenhuma assinatura ativa.
    post:
      tags:
        - Cliente & Painel
      summary: Cancelar ou reativar assinatura
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - action
              properties:
                action:
                  type: string
                  enum: [cancel, reactivate]
                  example: cancel
      responses:
        '200':
          description: Assinatura atualizada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assinatura'

  # --- CONTEÚDO & INVESTIGAÇÃO ---
...
components:
  schemas:
    Cliente:
      type: object
      properties:
        id:
          type: string
          example: cust-001
        nome:
          type: string
          example: Mariana Silva
        email:
          type: string
          format: email
          example: mariana.silva@email.com
        telefone:
          type: string
          example: (11) 98765-4321
        preferencias:
          $ref: '#/components/schemas/PreferenciasAssinante'
    PreferenciasAssinante:
      type: object
      properties:
        shirtSize:
          type: string
          example: M
        shoeSize:
          type: string
          example: 38
        notes:
          type: string
          example: Prefere tons escuros nas peças.
...
    Assinatura:
      type: object
      properties:
        id:
          type: string
          example: sub-001
        customerId:
          type: string
          example: cust-001
        planId:
          type: string
          example: plan-monthly
        planName:
          type: string
          example: Plano Mensal
        status:
          type: string
          enum: [active, pending_payment, cancelled, paused, past_due]
          example: active
        startedAt:
          type: string
          format: date-time
        nextBillingDate:
          type: string
          format: date-time
        nextBillingAmount:
          type: integer
          example: 14990
        currentCycleBoxId:
          type: string
        currentCycleBoxName:
          type: string
        canCancel:
          type: boolean
          example: true
        canReactivate:
          type: boolean
          example: false
...
    Erro:
      type: object
      properties:
        mensagem:
          type: string
          example: Credenciais inválidas.